[FEATURE] Introduce Action Directive
An Action Directive does not return a node but does an action on the context
